requirements.txt
aiogram==3.17.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
loguru==0.7.3
redis==5.2.1
pytest==8.3.5